Add forgotten "or later" to LGPLv2.1-licensed mods
This is NOT a change of the license terms, just an addition of forgotten information.

The "or later" was forgotten to be mentioned explicitly.
Repixture as a whole was already licensed under "LGPLv2.1 or later"
in version 3.14.0 and previous versions before, it just wasn't mentioned
properly in the individual mods.
See the LICENSE.txt file of that version.